+++ apps/mobile/src/components/common/StatusBar.tsx
