+++ apps/mobile/src/components/home/Timeline.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Text, View } from 'react-native'
+
+const Timeline = () => {
+  return (
+    <View>
+      <Text>asd</Text>
+    </View>
+  )
+}
+
+export default Timeline
